chatbot/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
gunicorn==21.2.0
gspread==5.12.0
oauth2client==4.1.3
pdfplumber==0.11.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3